Complete renaming
Complete renaming of classes, a lot of methods.
Also, at lot of small changes were made to the comments; but they need to be comed, because that are small naming-reference-mismatches for sure.
Also added is a speed test for instance methods versus CallbackMethods, versus CallMes. The results are confusing. Please run over 400 test ROUNDS to be sure to achieve stable results; at varying number of test rounds are requred before stable test time are achieved. Which is confusing. But once they are they, it good and stable, especially, wierdly ideally, in the range of ~25-120 tests per round.